components: renaming PresentationalComponent to connect
