+++ src/index.js
@@ -1,4 +1,4 @@
-import connect from './components/PresentationalComponent';
+import connect from './components/connect';
 import Provider from './components/Provider';
 
 export { connect, Provider };
